requirements.txt
dspy-ai>=3.0.3
datasets>=4.1.0
openai>=1.107.3
huggingface-hub>=0.34.5
requests>=2.31.0
